Implementation of overall changes to single module
